Continents written to ai_strategy, visualisation fixes
- Force countrydetail regeneration if any change is done to states
- Write continents in ai_areas and ai_strategy files
- Improve performance in strategic region center checks
- Update submodule Fwg

+++ RandomParadox/include/generic/ScenarioGenerator.h
@@ -42,7 +42,7 @@ public:
   double worldPopulationFactor = 1.0;
   double worldIndustryFactor = 1.0;
   double resourceFactor = 1.0;
-  double strategicRegionFactor = 1.0;
+  float strategicRegionFactor = 1.0;
   // containers - used for every game
   std::vector<ScenarioContinent> scenContinents;
   std::vector<std::shared_ptr<Region>> gameRegions;
